cleanup Term: unused stuff, comments, type-correctness, C++11 unions
