+++ Kernel/Term.cpp
@@ -15,27 +15,8 @@
  * @since 06/05/2007 Manchester, changed into a single class instead of three
  */
 
-#include <ostream>
-
-#include "Debug/Tracer.hpp"
-
-#include "Lib/Allocator.hpp"
-#include "Lib/Environment.hpp"
-#include "Lib/Portability.hpp"
-#include "Lib/Stack.hpp"
-#include "Lib/Set.hpp"
-#include "Lib/Int.hpp"
-#include "Lib/STL.hpp"
-
 #include "Indexing/TermSharing.hpp"
 
-#include "Shell/Options.hpp"
-#include "Shell/Statistics.hpp"
-
-#include "Formula.hpp"
-#include "Signature.hpp"
-#include "SortHelper.hpp"
-#include "Substitution.hpp"
 #include "SubstHelper.hpp"
 #include "TermIterators.hpp"
 #include "RobSubstitution.hpp"
@@ -43,12 +24,6 @@
 #include "Term.hpp"
 #include "FormulaVarIterator.hpp"
 
-/** If non-zero, term ite functors will be always expanded to
- * the ( p ? x : y ) notation on output */
-#define ALWAYS_OUTPUT_TERM_ITE 0
-
-
-using namespace std;
 using namespace Lib;
 using namespace Kernel;
 
@@ -65,7 +40,7 @@ const unsigned Term::SPECIAL_FUNCTOR_LOWER_BOUND;
 void* Term::operator new(size_t,unsigned arity, size_t preData)
 {
   CALL("Term::new");
-  //preData must be a multiply of pointer size to maintain alignment
+  //preData must be a multiple of pointer size to maintain alignment
   ASS_EQ(preData%sizeof(size_t), 0);
 
   size_t sz = sizeof(Term)+arity*sizeof(TermList)+preData;
@@ -472,35 +447,6 @@ bool Term::containsAllVariablesOf(Term* t)
   return true;
 }
 
-bool TermList::containsAllVariableOccurrencesOf(TermList t)
-{
-  CALL("TermList:containsAllVariableOccurrencesOf");
-  // varBalance[x] = (#occurrences of x in this) - (#occurrences of x in t)
-  static vunordered_map<unsigned int, int> varBalance(16);
-  varBalance.clear();
-
-  static VariableIterator vit;
-
-  // collect own vars
-  vit.reset(*this);
-  while (vit.hasNext()) {
-    int& bal = varBalance[vit.next().content()];
-    bal += 1;
-  }
-
-  // check that collected vars do not occur more often in t
-  vit.reset(t);
-  while (vit.hasNext()) {
-    int& bal = varBalance[vit.next().content()];
-    bal -= 1;
-    if (bal < 0) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
 bool Term::isShallow() const
 {
   CALL("Term::isShallow");
@@ -781,16 +727,8 @@ vstring Term::toString(bool topLevel) const
       res += arg1.toString(false) + " > " + arg2.toString();
       res += topLevel ? "" : ")";
       return res;
-    }/*else if(isApplication()){
-      ASS(arity() == 4);
-      vstring res;
-      TermList arg1 = *(nthArgument(2));
-      TermList arg2 = *(nthArgument(3));
-    //  res += topLevel ? "" : "("; 
-      res += "(" + arg1.toString() + " @ " + arg2.toString(false) + ")";
-    //  res += topLevel ? "" : ")";
-      return res;
-    }*/
+    }
+
     printArgs = isSort() || env.signature->getFunction(_functor)->combinator() == Signature::NOT_COMB;
   }
 
@@ -1151,7 +1089,7 @@ Term* Term::createLet(unsigned functor, VList* variables, TermList binding, Term
   s->getSpecialData()->_letData.functor = functor;
   s->getSpecialData()->_letData.variables = variables;
   s->getSpecialData()->_letData.sort = bodySort;
-  s->getSpecialData()->_letData.binding = binding.content();
+  s->getSpecialData()->_letData.binding = binding;
   return s;
 }
 
@@ -1191,7 +1129,7 @@ Term* Term::createTupleLet(unsigned tupleFunctor, VList* symbols, TermList bindi
   s->getSpecialData()->_letTupleData.functor = tupleFunctor;
   s->getSpecialData()->_letTupleData.symbols = symbols;
   s->getSpecialData()->_letTupleData.sort = bodySort;
-  s->getSpecialData()->_letTupleData.binding = binding.content();
+  s->getSpecialData()->_letTupleData.binding = binding;
   return s;
 } 
 
@@ -1903,8 +1841,6 @@ AtomicSort::AtomicSort()
   CALL("AtomicSort::AtomicSort/0");
 }
 
-#include <iostream>
-
 #if VDEBUG
 vstring Term::headerToString() const
 {
@@ -1970,50 +1906,6 @@ bool operator<(const TermList& lhs, const TermList& rhs)
     return lhs.var() < rhs.var();
   }
 }
-
-/**
- * If the literal has the form p(R,f(S),T), where f(S) is the
- * n-th argument, then return the literal, then return the
- * literal p%f(R,S,T).
- *//* 
-Literal* Literal::flattenOnArgument(const Literal* lit,int n)
-{
-  ASS(lit->shared());
-
-  const TermList* ts = lit->nthArgument(n);
-  ASS(! ts->isVar());
-  const Term* t = ts->term();
-  unsigned newArity = lit->arity() + t->arity() - 1;
-  vstring newName = lit->predicateName() + '%' + Int::toString(n) +
-                   '%' + t->functionName();
-  unsigned newPredicate = env.signature->addPredicate(newName,newArity);
-
-  Literal* newLiteral = new(newArity) Literal(newPredicate,newArity,
-					      lit->polarity(),false);
-  // copy all arguments
-  TermList* newArgs = newLiteral->args();
-  const TermList* args = lit->args();
-  for (int i = 0;i < n;i++) {
-    *newArgs = *args;
-    newArgs = newArgs->next();
-    args = args->next();
-  }
-  // now copy the arguments of t
-  for (const TermList* ss=t->args();! ss->isEmpty();ss = ss->next()) {
-    *newArgs = *ss;
-    newArgs = newArgs->next();
-  }
-  args = args->next();
-  while (! args->isEmpty()) {
-    *newArgs = *args;
-    newArgs = newArgs->next();
-    args = args->next();
-  }
-  ASS(newArgs->isEmpty());
-
-  return env.sharing->insert(newLiteral);
-} // Literal::flattenOnArgument
-*/
 
 bool Kernel::positionIn(TermList& subterm,TermList* term,vstring& position)
 {
